% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baixar_cpopg_tjac.R
\name{baixar_cpopg_tjac}
\alias{baixar_cpopg_tjac}
\title{Baixa consulta processual de primeiro grau}
\usage{
baixar_cpopg_tjac(processos = NULL, diretorio = ".")
}
\arguments{
\item{processos}{Processos a serem baixados}

\item{diretorio}{Diretório onde baixar os htmls.
Se não informado, baixa no atual.}
}
\value{
htmls
}
\description{
Baixa consulta processual de primeiro grau
}
\examples{
\dontrun{
baixar_cpopg_tjac("0004807-95.2019.8.01.0001")
}
}
