% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ler_dados_cpopg_tjac.R
\name{ler_dados_cpopg_tjac}
\alias{ler_dados_cpopg_tjac}
\title{lê dados do cpopg}
\usage{
ler_dados_cpopg_tjac(arquivos = NULL, diretorio = ".")
}
\arguments{
\item{arquivos}{Caminho dos arquivos}

\item{diretorio}{Se arquivos = NULL, informar diretório}
}
\value{
tibble
}
\description{
lê dados do cpopg
}
