% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ler_partes_tjac.R
\name{ler_partes_tjac}
\alias{ler_partes_tjac}
\title{Aplica parser para extrair informações sobre as partes do
    processo.}
\usage{
ler_partes_tjac(arquivos = NULL, diretorio = ".")
}
\arguments{
\item{arquivos}{O diretório é ignorado se você fornecer o vetor
de arquivos}

\item{diretorio}{Objeto ou diretório onde se encontram os htmls baixados.}
}
\value{
tabela com informações das partes.
}
\description{
Aplica parser para extrair informações sobre as partes do
    processo.
}
