% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ler_tabela_documentos_tjac.R
\name{ler_tabela_documentos_tjac}
\alias{ler_tabela_documentos_tjac}
\title{lê lista de documentos}
\usage{
ler_tabela_documentos_tjac(arquivos = NULL, diretorio = ".")
}
\arguments{
\item{arquivos}{Caminho dos arquivos. Se NULL, informar diretório.}

\item{diretorio}{Informar se arquivos = NULL.}
}
\value{
tibble
}
\description{
lê lista de documentos
}
\examples{
\dontrun{
docs <- ler_tabela_documentos()
}
}
